requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.0
MarkupSafe==2.1.1
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
tinydb==4.7.0
twilio==7.8.0
urllib3==1.26.9
Werkzeug==2.0.3
